folder structure refactor (#2247)
