+++ app/src/features/apiClient/routes.tsx
@@ -3,7 +3,7 @@ import PATHS from "config/constants/sub/paths";
 import ApiClientFeatureContainer from "./container";
 import { APIClient } from "./screens/apiClient/APIClient";
 import ProtectedRoute from "components/authentication/ProtectedRoute";
-import { VariablesListView } from "./screens/apiClient/components/variablesListView/VariablesListView";
+import { EnvironmentView } from "./screens/environment/environmentView";
 
 export const apiClientRoutes: RouteObject[] = [
   {
@@ -24,11 +24,11 @@ export const apiClientRoutes: RouteObject[] = [
       },
       {
         path: PATHS.API_CLIENT.ENVIRONMENTS.INDEX + "/:envId",
-        element: <ProtectedRoute component={VariablesListView} />,
+        element: <ProtectedRoute component={EnvironmentView} />,
       },
       {
         path: PATHS.API_CLIENT.ENVIRONMENTS.NEW.RELATIVE,
-        element: <ProtectedRoute component={VariablesListView} />,
+        element: <ProtectedRoute component={EnvironmentView} />,
       },
     ],
   },
