+++ app/src/features/apiClient/screens/environment/components/VariablesList/components/customTableRow/CustomTableRow.tsx
