+++ app/src/features/apiClient/screens/environment/components/VariablesList/hooks/useVariablesListColumns.tsx
