+++ app/src/features/apiClient/screens/environment/components/VariablesList/variablesList.scss
