+++ app/src/features/apiClient/screens/environment/components/VariablesListHeader/variablesListHeader.scss
